+++ example.py
@@ -10,7 +10,6 @@ import os
 import aiohttp
 from dotenv import load_dotenv
 from loguru import logger
-
 from pipecat.audio.turn.smart_turn.base_smart_turn import SmartTurnParams
 from pipecat.audio.turn.smart_turn.local_smart_turn_v3 import LocalSmartTurnAnalyzerV3
 from pipecat.audio.vad.silero import SileroVADAnalyzer